feat: creating navbar-scroll.service
